python-projects/nosql-hospital-services/requirements.txt
#Mongo
falcon
pymongo
uvicorn
requests

#Chroma
chromadb
transformers
sentence-transformers

#Cassandra
cassandra-driver
time_uuid

#Dgraph
pydgraph